Moved rename_tb_link to utils.py
Moved rename_tb_links to utils.py since it makes
more sense to keep it there.

Also fixed some flake8 errors.
