+++ pbench/server/__init__.py
@@ -298,27 +298,3 @@ def md5sum(filename):
         for buf in iter(partial(f.read, 128), b""):
             d.update(buf)
     return d.hexdigest()
-
-
-def rename_tb_link(tb, dest, logger):
-    try:
-        os.mkdir(dest)
-    except FileExistsError:
-        # directory already exists, ignore
-        pass
-    except Exception:
-        logger.exception(
-            "os.mkdir: Unable to create tar ball destination directory: {}".format(dest)
-        )
-        raise
-    tbname = os.path.basename(tb)
-    tbnewname = os.path.join(dest, tbname)
-    try:
-        os.rename(tb, tbnewname)
-    except Exception:
-        logger.exception(
-            "os.rename: Unable to move tar ball link {} to destination directory: {}".format(
-                tb, dest
-            )
-        )
-        raise
